registeration

+++ base/views.py
@@ -9,6 +9,8 @@ from django.contrib import messages
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
+from django.contrib.auth.forms import UserCreationForm
 # Create your views here.
 # rooms = [
 #   {'id': 1, 'name': 'lets learn python'},
@@ -17,8 +19,12 @@ from django.contrib.auth.decorators import login_required
 # ]
 
 def loginPage(request):
+  page = 'login'
+  if request.user.is_authenticated:
+    return redirect('home')
+
   if request.method == 'POST':
-    username = request.POST.get('username')
+    username = request.POST.get('username').lower()
     password = request.POST.get('password')
 
     try:
@@ -33,13 +39,29 @@ def loginPage(request):
       return redirect('home')
     else:
       messages.error(request, 'username or password is wrong')
-  context = {}
+  context = {'page': page}
   return render(request, 'base/login_register.html', context)
 
 
 def logoutUser(request):
   logout(request)
   return redirect('home')
+
+def registerPage(request):
+  form = UserCreationForm
+
+  if request.method == 'POST':
+    form = UserCreationForm(request.POST)
+    if form.is_valid():
+      user = form.save(commit=False)
+      user.username = user.username.lower()
+      user.save()
+      login(request, user)
+      return redirect('home')
+    else:
+      messages.error(request, 'an error has occured')
+
+  return render(request, 'base/login_register.html', {'form': form})
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -76,6 +98,9 @@ def updateRoom(request, pk):
   room = Room.objects.get(id=pk)
   form = RoomForm(instance=room)
 
+  if request.user != room.host:
+    return HttpResponse('You are not allowed here')
+
   if request.method == 'POST':
     form = RoomForm(request.POST, instance=room)
     if form.is_valid():
@@ -87,6 +112,10 @@ def updateRoom(request, pk):
 @login_required(login_url='login')
 def deleteRoom(request, pk):
   room = Room.objects.get(id=pk)
+
+  if request.user != room.host:
+    return HttpResponse('You are not allowed here')
+  
   if request.method == 'POST':
     room.delete()
     return redirect('home')
